% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__aT.x.DvecSigma4beta.inv.R
\name{aT.x.DvecSigma4beta.inv}
\alias{aT.x.DvecSigma4beta.inv}
\title{<description>}
\usage{
aT.x.DvecSigma4beta.inv(
  aT,
  Sigma.inv,
  diag.K.list,
  Xmats.delta.knots,
  P.type,
  p,
  q,
  q.i,
  n.par4knots
)
}
\arguments{
\item{aT}{NA}

\item{Sigma.inv}{NA}

\item{diag.K.list}{NA}

\item{Xmats.delta.knots}{NA}

\item{P.type}{NA}

\item{p}{NA}

\item{q}{NA}

\item{q.i}{NA}

\item{n.par4knots}{NA}
}
\value{
NA
}
\description{
<details>
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
