% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__convert.densParams.R
\name{convert.densParams}
\alias{convert.densParams}
\title{Convert the parameters corresponding to different transformations}
\usage{
convert.densParams(mean, var, linkage)
}
\arguments{
\item{mean}{NA}

\item{var}{NA}

\item{linkage}{NA}
}
\value{
NA
}
\description{
Details.
}
\note{
First version: Wed Feb 02 16:09:43 CET 2011;
      Current: Wed Feb 02 16:09:48 CET 2011.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
