% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__delta.sumlogdetP.R
\name{delta.sumlogdetP}
\alias{delta.sumlogdetP}
\title{<description>}
\usage{
delta.sumlogdetP(P.mats, q.i, Xmats.delta.knots.lst, priorArgs, n.par4knots)
}
\arguments{
\item{P.mats}{NA}

\item{q.i}{NA}

\item{Xmats.delta.knots.lst}{NA}

\item{priorArgs}{NA}

\item{n.par4knots}{NA}
}
\value{
NA
}
\description{
<details>
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
