% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__delta.sumqlogdetK.R
\name{delta.sumqlogdetK}
\alias{delta.sumqlogdetK}
\title{gradient for sum q_i log|K_i|}
\usage{
delta.sumqlogdetK(q.i, diag.K)
}
\arguments{
\item{q.i}{NA}

\item{diag.K}{NA}

\item{args}{"list" args$subset: the subsets of the diagonal K matrix.}
}
\value{
"matrix" 1-by-
}
\description{
Details
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
