% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/models__linear__etc__delta.vecPartiSigma4beta.inv.R
\name{delta.vecPartiSigma4beta.inv}
\alias{delta.vecPartiSigma4beta.inv}
\title{This is the gradient for a single partition of beta's variance.}
\usage{
delta.vecPartiSigma4beta.inv(diag.K_i, Sigma.inv, P_i)
}
\arguments{
\item{diag.K_i}{NA}

\item{Sigma.inv}{NA}

\item{P_i}{NA}
}
\value{
"matrix"
}
\description{
Details from the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
