% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__deriv_prior.R
\name{deriv_prior}
\alias{deriv_prior}
\title{A collection of gradient for common priors.}
\usage{
deriv_prior(B, priorArgs, hessMethod)
}
\arguments{
\item{B}{"matrix".  The paramter that need to be added with a prior. The gradient and
hessian are calculated conditional on B. B should be always an one-column matrix,}

\item{priorArgs}{"list".  priorArgs$prior_type: when prior_type is set to "mvnorm",
you have to provide priorArgs$mean: "matrix", the mean of parameter, mu0 should be
always an one-column matrix; priorArgs$covariance: "matrix", the covariance
matrix. A g-prior can be constructed by setting it to X'X, where X is the
covariates matrix.; priorArgs$shrinkage: "numeric", the shrinkage for the
covariance.}

\item{hessMethod}{NA}
}
\value{
"list". The gradient and hessian matrix, see below.  `gradObsPri`
    "matrix". One-colunm. `hessObsPri` "matrix". A squre matrix. Dimension same as
    prior_type$Sigma0.
}
\description{
The parameters after "..." should be matched exactly.
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
