% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp__DGP.hwang.R
\name{DGP.hwang}
\alias{DGP.hwang}
\title{A collection of DGPs for different models from Hwang's example.}
\usage{
DGP.hwang(
  n,
  q,
  Sigma,
  model,
  otherArgs = list(seed = NA, nTesting = NA),
  PlotData
)
}
\arguments{
\item{n}{"numeric".  No. of obs. to be generated}

\item{q}{NA}

\item{Sigma}{"matrix".  The variance covariance matrix.}

\item{model}{"character".  The model name to be use.  When the model is "spline", you
also need to supply paramters, see otherArgs.}

\item{otherArgs}{"list".  Other arguemnts need to pass to the function.
otherArgs$seed: Set the seed. If this vaule is unset, use a random seed.
otherArgs$xi: otherArgs$withOnt: otherArgs$spline.type:}

\item{PlotData}{"logical" If need to plot the data out.}
}
\value{
"list"
}
\description{
Data generating processes. The model can be extended.
}
\references{
Hwang 2000
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
