% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp__DGP.surface.R
\name{DGP.surface}
\alias{DGP.surface}
\title{DGP Surface nested}
\usage{
DGP.surface(n, p, q.o, q.s, Sigma, splineArgs, splineArgs.crl, PlotDGP = FALSE)
}
\arguments{
\item{n}{NA}

\item{p}{NA}

\item{q.o}{NA}

\item{q.s}{NA}

\item{Sigma}{NA}

\item{splineArgs}{NA}

\item{splineArgs.crl}{NA}

\item{PlotDGP}{NA}
}
\value{
NA
}
\description{
A DGP process where the true model nests in the fitted model
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
