% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__FitDiagnosis.R
\name{FitDiagnosis}
\alias{FitDiagnosis}
\title{Diagnosis if the spline model}
\usage{
FitDiagnosis(
  x,
  Y,
  OUT.Params,
  Data.gen,
  logpost.fun.name,
  splineArgs,
  splineArgs.gen,
  Params_Transform,
  crossvalid.struc,
  burn.in,
  criterion = c("LOSS", "KL", "L2")
)
}
\arguments{
\item{x}{NA}

\item{Y}{NA}

\item{OUT.Params}{NA}

\item{Data.gen}{NA}

\item{logpost.fun.name}{NA}

\item{splineArgs}{NA}

\item{splineArgs.gen}{NA}

\item{Params_Transform}{NA}

\item{crossvalid.struc}{NA}

\item{burn.in}{NA}

\item{criterion}{NA}
}
\value{
NA
}
\description{
This function is only applied to the moving knots model.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
