% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__glm__glm_gradhess.R
\name{glm_gradhess}
\alias{glm_gradhess}
\title{GLM}
\usage{
glm_gradhess(
  Params,
  hessMethod,
  Y,
  x,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{Params}{NA}

\item{hessMethod}{NA}

\item{Y}{NA}

\item{x}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\description{
GLM
}
