% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__grad.x.deriv_link.R
\name{grad.x.deriv_link}
\alias{grad.x.deriv_link}
\title{Gradient derivative for x wrt link function}
\usage{
grad.x.deriv_link(grad, Param, link)
}
\arguments{
\item{grad}{"matrix" gradient}

\item{Param}{"matrix" should always be a matrix}

\item{link}{NA}
}
\value{
NA
}
\description{
Just the chain rule.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
