% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__glm__glm_gradhess.R
\name{grad_vech_Sigma}
\alias{grad_vech_Sigma}
\title{The gradient with respect to vech Sigma}
\usage{
grad_vech_Sigma(B, Sigma, x, xi, l0, l, link, gradient.prior.Sigma)
}
\arguments{
\item{B}{NA}

\item{Sigma}{NA}

\item{x}{NA}

\item{xi}{NA}

\item{l0}{NA}

\item{l}{NA}

\item{link}{NA}

\item{gradient.prior.Sigma}{NA}

\item{gradient.prior.vecB}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{Y}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\description{
sid
}
