% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__glm__glm_gradhess.R
\name{gradient_vecB}
\alias{gradient_vecB}
\title{Gradient w.r.t. vecB}
\usage{
gradient_vecB(B, Sigma, x, xi, l0, l, link, gradient.prior.vecB)
}
\arguments{
\item{B}{NA}

\item{Sigma}{NA}

\item{x}{NA}

\item{xi}{NA}

\item{l0}{NA}

\item{l}{NA}

\item{link}{NA}

\item{gradient.prior.vecB}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{Y}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\description{
Be aware that this will give a q--by--1 matrix
}
