% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__glm__glm_gradhess.R
\name{gradient_xi}
\alias{gradient_xi}
\title{gradient w.r.t. xi}
\usage{
gradient_xi(Y, x, xi, l0, l, n0, S0, B, ka, gradient.prior.xi)
}
\arguments{
\item{Y}{NA}

\item{x}{NA}

\item{xi}{NA}

\item{l0}{NA}

\item{l}{NA}

\item{n0}{NA}

\item{S0}{NA}

\item{B}{NA}

\item{ka}{NA}

\item{gradient.prior.xi}{NA}

\item{Sigma}{NA}

\item{link}{NA}

\item{gradient.prior.Sigma}{NA}

\item{gradient.prior.vecB}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\description{
gradient w.r.t. xi
}
