% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__glm__glm_gradhess.R
\name{gradient_xi_condi}
\alias{gradient_xi_condi}
\title{Gradient w.r.t xi (conditional method)}
\usage{
gradient_xi_condi(B, Sigma, x, xi, l0, l, link, gradient.prior.xi)
}
\arguments{
\item{B}{NA}

\item{Sigma}{NA}

\item{x}{NA}

\item{xi}{NA}

\item{l0}{NA}

\item{l}{NA}

\item{link}{NA}

\item{gradient.prior.xi}{NA}

\item{Y}{NA}

\item{n0}{NA}

\item{S0}{NA}

\item{ka}{NA}

\item{gradient.prior.Sigma}{NA}

\item{gradient.prior.vecB}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\description{
Be aware that this will give k--by-1 matrix
}
