% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__idx.b2beta.R
\name{idx.b2beta}
\alias{idx.b2beta}
\title{Make indices from b to beta.}
\usage{
idx.b2beta(p, q, q.i)
}
\arguments{
\item{p}{NA}

\item{q}{NA}

\item{q.i}{NA}
}
\value{
NA
}
\description{
Details from the paper
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
