% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__idx.beta2b.R
\name{idx.beta2b}
\alias{idx.beta2b}
\title{Convert indices from beta to b}
\usage{
idx.beta2b(p, q, q.i)
}
\arguments{
\item{p}{NA}

\item{q}{NA}

\item{q.i}{NA}
}
\value{
NA
}
\description{
Details from the paper
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
