% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__knots_check_boundary.R
\name{knots_check_boundary}
\alias{knots_check_boundary}
\title{Boundary check for the moving knots model
It is easy to check with in 2D. But how to manage it in >3D space.  Check if weather
each dim if fit.  The usual way shoud be check if new convex hull is same as the old
convex hull by adding new data in. But that is always time comsumming. But if we know
the boundary of each dimmesion, that can be easier.}
\usage{
knots_check_boundary(P4X, method = "singular")
}
\arguments{
\item{P4X}{NA}

\item{method}{"character"}

\item{...}{Input with respect to different methods.}
}
\value{
"character" "ok": the knots are at good position. "bad": bad knots locations.
}
\description{
Boundary check for the moving knots model
It is easy to check with in 2D. But how to manage it in >3D space.  Check if weather
each dim if fit.  The usual way shoud be check if new convex hull is same as the old
convex hull by adding new data in. But that is always time comsumming. But if we know
the boundary of each dimmesion, that can be easier.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
