% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__knots_mat2list.R
\name{knots_mat2list}
\alias{knots_mat2list}
\title{Convert the knots matrix into the knots list}
\usage{
knots_mat2list(knots.mat, splineArgs)
}
\arguments{
\item{knots.mat}{NA}

\item{splineArgs}{NA}
}
\value{
NA
}
\description{
Details available from the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
