% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__knots_subsetsIdx.R
\name{knots_subsetsIdx}
\alias{knots_subsetsIdx}
\title{Find the locations of subsets in the knots matrix}
\usage{
knots_subsetsIdx(knots.subsets, k, m)
}
\arguments{
\item{knots.subsets}{"list"}

\item{k}{"integer"
No. of knots in the knots matrix.}

\item{m}{"integer"
Dimension of knots in the knots matrix.}
}
\value{
"list"
}
\description{
Details see the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
