% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__KStepNewtonMove.R
\name{KStepNewtonMove}
\alias{KStepNewtonMove}
\title{Newton move for spline models without dimession changes.}
\usage{
KStepNewtonMove(
  param.cur,
  gradhess.fun.name,
  KStep,
  Params,
  hessMethod,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{param.cur}{"matrix".  The initial values for the Newton update.}

\item{gradhess.fun.name}{NA}

\item{KStep}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\value{
"list". See bellow.  `gradObs.cur`: The gradient matrix. `invHessObs.cur`: The
    inverse hessian matrix.  `param.cur`: The updated parameters after K step Newton
    iterations.
}
\description{
Details see the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
