% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__linear_IWishart.R
\name{linear_IWishart}
\alias{linear_IWishart}
\title{Calculate the posterior df and location matrix V from the inverse Wishart distribution.}
\usage{
linear_IWishart(Params, Y, x0, splineArgs, priorArgs, Params_Transform)
}
\arguments{
\item{Params}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
