% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__linear_logpost.R
\name{linear_logpost}
\alias{linear_logpost}
\title{The conditional and joint log posterior function}
\usage{
linear_logpost(
  Y,
  x0,
  Params,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{Y}{NA}

\item{x0}{NA}

\item{Params}{NA
Params$xi:
Parmas$K:}

\item{callParam}{callParam$id: It should be able to obtain conditianl posterior or joint
posterior or just likelihood.
callParam$subset: If the paramater set is larg, we may only update a subset of
them.}

\item{splineArgs}{"list".}

\item{priorArgs}{priorArgs$prior_type:
priorArgs$n0:
priorArgs$S0:
priorArgs$mu:
priorArgs$M:
priorArgs$mu0:
priorArgs$Sigma0:
priorArgs$ka0:
priorArgs$ka.mu0:
priorArgs$ka.Sigma0:
priorArgs$Sigma.mu0:
priorArgs$Sigma.Sigma0:}

\item{Params_Transform}{NA}

\item{ParamsTransArgs}{"list"}
}
\value{
"scalar".
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
