% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__LogPredScore.R
\name{LogPredScore}
\alias{LogPredScore}
\title{Log predictive scores}
\usage{
LogPredScore(
  Y,
  x,
  logpost.fun.name,
  crossvaid.struc,
  splineArgs,
  priorArgs,
  OUT.Params,
  Params_Transform,
  burn.in,
  LPDS.sampleProp
)
}
\arguments{
\item{Y}{NA}

\item{x}{NA}

\item{logpost.fun.name}{NA}

\item{crossvaid.struc}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{OUT.Params}{NA}

\item{Params_Transform}{NA}

\item{burn.in}{NA}

\item{LPDS.sampleProp}{NA}
}
\value{
NA
}
\description{
Details see the paper
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
