% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__make.knotsPriVar.R
\name{make.knotsPriVar}
\alias{make.knotsPriVar}
\title{Set the prior variance of the knots.}
\usage{
make.knotsPriVar(x, splineArgs)
}
\arguments{
\item{x}{NA}

\item{splineArgs}{NA}
}
\value{
NA
}
\description{
Details from the paper
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
