% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__Mat.x.delta.xi.R
\name{Mat.delta.xi}
\alias{Mat.delta.xi}
\title{<description>}
\usage{
Mat.delta.xi(Mat, delta.knots, n, p, q.i)
}
\arguments{
\item{Mat}{NA}

\item{delta.knots}{NA}

\item{n}{NA}

\item{p}{NA}

\item{q.i}{NA}
}
\value{
NA
}
\description{
<details>
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
