% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/models__linear__etc__Mat.x.AT.k.I.x.K.x.delta.knots.R
\name{Mat.x.AT.k.I.x.K.x.delta.knots}
\alias{Mat.x.AT.k.I.x.K.x.delta.knots}
\title{<description>}
\usage{
Mat.x.AT.k.I.x.K.x.delta.knots(Mat, A, delta.knots, p, q, q.i)
}
\arguments{
\item{Mat}{"matrix" Sigma4beta.tilde}

\item{A}{NA}

\item{delta.knots}{NA}

\item{p}{NA}

\item{q}{NA}

\item{q.i}{NA}
}
\value{
NA
}
\description{
<details>
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
