% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/models__linear__etc__Mat.x.delta.vecSigma4beta.inv.R
\name{Mat.x.delta.vecSigma4beta.inv}
\alias{Mat.x.delta.vecSigma4beta.inv}
\title{Compute a matrix to pre-multiply the gradient for vec beta's inverse variance}
\usage{
Mat.x.delta.vecSigma4beta.inv(Mat, delta4K.list, p, q, q.i)
}
\arguments{
\item{Mat}{NA}

\item{delta4K.list}{NA}

\item{p}{NA}

\item{q}{NA}

\item{q.i}{NA}

\item{args}{"list" args$subset: subset of the shrinkage}
}
\value{
NA
}
\description{
<details>
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
