% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__Mat.x.DvecA.k.XTX_stp1.R
\name{Mat.x.DvecA.k.XTX_stp1}
\alias{Mat.x.DvecA.k.XTX_stp1}
\title{Perform a dense matrix multiplying by Dev[vec[A %*% P_i]]/Dev[vecX_i'] %*% (I_q_i^2 +
K_qi,qi)}
\usage{
Mat.x.DvecA.k.XTX_stp1(Mat, A, p, q)
}
\arguments{
\item{Mat}{"Matrix"}

\item{A}{"Matrix"}

\item{p}{NA}

\item{q}{NA}
}
\value{
"Matrix"
}
\description{
Part 1.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
