% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__Mat.x.DvecA.k.XTX_stp2.R
\name{Mat.x.DvecA.k.XTX_stp2}
\alias{Mat.x.DvecA.k.XTX_stp2}
\title{<description>}
\usage{
Mat.x.DvecA.k.XTX_stp2(Mat, X.delta.knots.i, p, q, q_i)
}
\arguments{
\item{Mat}{"matrix"}

\item{X.delta.knots.i}{NA}

\item{p}{NA}

\item{q}{NA}

\item{q_i}{NA}

\item{X_i}{"matrix" the matrix used for obtaining P_i matrix}

\item{delta_i}{"matrix" the matrix for the gradient}
}
\value{
"matrix"
}
\description{
<details>
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
