% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/models__linear__etc__Mat.x.DvecSigma.inv.k.XTX.R
\name{Mat.x.DvecSigma.inv.k.XTX}
\alias{Mat.x.DvecSigma.inv.k.XTX}
\title{<description>}
\usage{
Mat.x.DvecSigma.inv.k.XTX(Mat, Sigma.inv, X.delta.knots.lst, p, q, q.i)
}
\arguments{
\item{Mat}{NA}

\item{Sigma.inv}{NA}

\item{X.delta.knots.lst}{NA}

\item{p}{NA}

\item{q}{NA}

\item{q.i}{NA}
}
\value{
NA
}
\description{
<details>
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
