% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__MCMC.trajectory.R
\name{MCMC.trajectory}
\alias{MCMC.trajectory}
\title{Trajectory MCMC for movingknots}
\usage{
MCMC.trajectory(iIter, nIter, iCross, OUT.accept.probs, interval = 0.1)
}
\arguments{
\item{iIter}{NA}

\item{nIter}{NA}

\item{iCross}{NA}

\item{OUT.accept.probs}{NA}

\item{interval}{e.g. 0.1}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
