% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__MHPropMain.R
\name{MHPropMain}
\alias{MHPropMain}
\title{The Main MCMC algorithm for movingknots.}
\usage{
MHPropMain(
  param.cur,
  gradhess.fun.name,
  logpost.fun.name,
  nNewtonStep,
  Params,
  hessMethod,
  Y.iCross,
  x.iCross,
  callParam,
  splineArgs,
  priorArgs,
  algArgs,
  Params_Transform,
  propMethod
)
}
\arguments{
\item{param.cur}{NA}

\item{gradhess.fun.name}{NA}

\item{logpost.fun.name}{NA}

\item{nNewtonStep}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{Y.iCross}{NA}

\item{x.iCross}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{algArgs}{NA}

\item{Params_Transform}{NA}

\item{propMethod}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
