% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__MHPropWithIWishart.R
\name{MHPropWithIWishart}
\alias{MHPropWithIWishart}
\title{Random walk Metropolis–Hastings algorithm for Sigma}
\usage{
MHPropWithIWishart(
  param.cur,
  logpost.fun.name,
  Params,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{param.cur}{NA}

\item{logpost.fun.name}{NA}

\item{Params}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
