% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__MHPropWithKStepNewton.R
\name{MHPropWithKStepNewton}
\alias{MHPropWithKStepNewton}
\title{Metropolis–Hastings algorithm with K-step Newton method for the spline model.}
\usage{
MHPropWithKStepNewton(
  param.cur,
  gradhess.fun.name,
  logpost.fun.name,
  nNewtonStep,
  Params,
  hessMethod,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  prop.df,
  Params_Transform
)
}
\arguments{
\item{param.cur}{NA}

\item{gradhess.fun.name}{NA}

\item{logpost.fun.name}{NA}

\item{nNewtonStep}{NA}

\item{Params}{NA}

\item{hessMethod}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{prop.df}{NA}

\item{Params_Transform}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
