% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__MovingKnots_MCMC.R
\name{MovingKnots_MCMC}
\alias{MovingKnots_MCMC}
\title{MCMC for movingknots.}
\usage{
MovingKnots_MCMC(
  gradhess.fun.name,
  logpost.fun.name,
  nIter,
  Params,
  Params4Gibbs,
  Params.sub.struc,
  Y,
  x0,
  splineArgs,
  priorArgs,
  Params_Transform,
  propMethods,
  algArgs,
  crossvalid.struc,
  OUT.Params,
  OUT.accept.probs,
  burn.in,
  LPDS.sampleProp,
  track.MCMC
)
}
\arguments{
\item{gradhess.fun.name}{NA}

\item{logpost.fun.name}{NA}

\item{nIter}{NA}

\item{Params}{NA}

\item{Params4Gibbs}{NA}

\item{Params.sub.struc}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}

\item{propMethods}{NA}

\item{algArgs}{NA}

\item{crossvalid.struc}{NA}

\item{OUT.Params}{NA}

\item{OUT.accept.probs}{NA}

\item{burn.in}{NA}

\item{LPDS.sampleProp}{NA}

\item{track.MCMC}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
