% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__P.matrix.R
\name{P.matrix}
\alias{P.matrix}
\title{Generate the P matrices for the moving knots model}
\usage{
P.matrix(X, q.i, priorArgs)
}
\arguments{
\item{X}{"matrix" the design matrix generated by d.matrix() function.}

\item{q.i}{NA}

\item{priorArgs}{NA}

\item{q}{"numeric" the dimensions for the P matrices}

\item{args}{"list" args$P.type: "character" the type for the P matrices}
}
\value{
"list" Containing the P matrices.
}
\description{
P matrix. Details available in the paper.
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
