% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__par.transform.R
\name{par.transform}
\alias{par.transform}
\title{Parameter transformation.}
\usage{
par.transform(par, method)
}
\arguments{
\item{par}{NA}

\item{method}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\note{
First version: ; Current: .
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
