% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__par.transform2.R
\name{par.transform2}
\alias{par.transform2}
\title{Transform the parameters from the original scale to the new scale (the final scale) wrt the linkages}
\usage{
par.transform2(par, method)
}
\arguments{
\item{par}{NA}

\item{method}{NA}
}
\value{
NA
}
\description{
Reverse step of par.transfrom()
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
