% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__Params.subsets.R
\name{Params.subsets}
\alias{Params.subsets}
\title{Organize the subsets of the parameters by taking away the fixed parameters.}
\usage{
Params.subsets(p, splineArgs, Params_Fixed, Params_subsetsArgs)
}
\arguments{
\item{p}{NA}

\item{splineArgs}{NA}

\item{Params_Fixed}{NA}

\item{Params_subsetsArgs}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
