% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__SurfacePred.R
\name{PredSurface}
\alias{PredSurface}
\title{Surface prediction for the movingknots.}
\usage{
PredSurface(Y, x0, OUT.FITTED, jIter, iCross, splineArgs)
}
\arguments{
\item{Y}{NA}

\item{x0}{NA}

\item{OUT.FITTED}{NA}

\item{jIter}{NA}

\item{iCross}{NA}

\item{splineArgs}{NA}
}
\description{
Details are available in the paper.
}
