% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__RandomWalkMetropolis.R
\name{RandomWalkMetropolis}
\alias{RandomWalkMetropolis}
\title{Random walk Metropolis algorithm for movingknots.}
\usage{
RandomWalkMetropolis(
  param.cur,
  logpost.fun.name,
  Params,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{param.cur}{NA}

\item{logpost.fun.name}{NA}

\item{Params}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\value{
NA
}
\description{
Random Walk Metropolis
}
\references{
Li Villani 2012
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
