% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__Sigma4betaFun.R
\name{Sigma4betaFun}
\alias{Sigma4betaFun}
\title{Calculate the variance for the prior of coefficients (beta) in the moving knots model}
\usage{
Sigma4betaFun(diag.K, Sigma, P.mats, inverse = FALSE)
}
\arguments{
\item{diag.K}{"matrix" diagonal part of the matrix K}

\item{Sigma}{"matrix" variance for the error term}

\item{P.mats}{NA}

\item{inverse}{"logical" if TRUE,  will return the inverse variance matrix.}

\item{P}{"list" containing the P matrices.}
}
\value{
"matrix"
}
\description{
Details available in the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
