% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms__sub.hessian.R
\name{sub.hessian}
\alias{sub.hessian}
\title{Extract the subset of hessian matrix.}
\usage{
sub.hessian(hessian, subset)
}
\arguments{
\item{hessian}{NA}

\item{subset}{NA}
}
\description{
Details are available in the paper.
}
