% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp__surface.hwang.R
\name{surface.hwang}
\alias{surface.hwang}
\title{Surface from Hwang paper.}
\usage{
surface.hwang(model, X, ...)
}
\arguments{
\item{model}{NA}

\item{X}{NA}

\item{...}{NA}
}
\value{
NA
}
\description{
Details are available in the paper.
}
