% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__linear__etc__X.x.delta.xi.R
\name{X.x.delta.xi}
\alias{X.x.delta.xi}
\title{Preform X multiply Dev xi}
\usage{
X.x.delta.xi(X, delta.knots, q.i)
}
\arguments{
\item{X}{NA}

\item{delta.knots}{NA}

\item{q.i}{NA}
}
\value{
NA
}
\description{
<details>
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
